add mode support for keybinds
Every keybind defaults to normal mode. To insert text change to Insert
mode. If you need to passthrough the Escape key, change to Passthrough
mode.

Edit modeindicator to customize the character to be displayed in the window
title. Also be sure to add a new char everytime you create a new mode.

If you create a new mode that needs to process only the keybinds assigned to
it (and discard every non matching key), return TRUE in winevent function at
surf.c. You can see that Normal mode is also there, because it needs to
do exactly that.